// src/graphql/orders/queries/getOrder.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: { input: any; output: any; }
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: { input: any; output: any; }
  /**
   * The `GenericScalar` scalar type represents a generic
   * GraphQL scalar value that could be:
   * String, Boolean, Int, Float, List or Object.
   */
  GenericScalar: { input: any; output: any; }
  /**
   * Allows use of a JSON String for input / output from the GraphQL schema.
   *
   * Use of this type is *not recommended* as you lose the benefits of having a defined, static
   * schema (one of the key benefits of GraphQL).
   */
  JSONString: { input: any; output: any; }
  /**
   * The `Time` scalar type represents a Time value as
   * specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Time: { input: any; output: any; }
  /**
   * Create scalar that ignores normal serialization/deserialization, since
   * that will be handled by the multipart request spec
   */
  Upload: { input: any; output: any; }
};
// ...
export type CityNode = Node & {
  __typename?: 'CityNode';
  alternateNames?: Maybe<Scalars['String']['output']>;
  city: ServiceContactNodeConnection;
  cityModel: DownloadLogNodeConnection;
  contacts: ContactNodeConnection;
  country: CountryNode;
  displayName: Scalars['String']['output'];
  featureCode?: Maybe<Scalars['String']['output']>;
  geonameId?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  jobApplications: JobApplicationNodeConnection;
  jobpositionSet: JobPositionNodeConnection;
  latitude?: Maybe<Scalars['Float']['output']>;
  longitude?: Maybe<Scalars['Float']['output']>;
  medicalfilmsolutionSet: MedicalFilmSolutionNodeConnection;
  name: Scalars['String']['output'];
  nameAscii: Scalars['String']['output'];
  offices: OfficeNodeConnection;
  population?: Maybe<Scalars['Int']['output']>;
  profiles: ProfileNodeConnection;
  promotedprojectSet: PromotedProjectNodeConnection;
  region?: Maybe<RegionNode>;
  samplescontact: SampleContactNodeConnection;
  searchNames: Scalars['String']['output'];
  shippingAddresses: Array<ShippingAddressNode>;
  slug: Scalars['String']['output'];
  stores: StoreNodeConnection;
  timezone?: Maybe<Scalars['String']['output']>;
  vacancies: VacantNodeConnection;
};
// ...
export type CountryNode = Node & {
  __typename?: 'CountryNode';
  alternateNames?: Maybe<Scalars['String']['output']>;
  citySet: CityNodeConnection;
  code2?: Maybe<Scalars['String']['output']>;
  code3?: Maybe<Scalars['String']['output']>;
  continent?: Maybe<Scalars['String']['output']>;
  formproposalSet: FormProposalNodeConnection;
  geonameId?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  nameAscii: Scalars['String']['output'];
  offices?: Maybe<OfficeNodeConnection>;
  phone?: Maybe<Scalars['String']['output']>;
  projects: ProjectNodeConnection;
  promotedprojectSet: PromotedProjectNodeConnection;
  regionSet: RegionNodeConnection;
  slug: Scalars['String']['output'];
  tld: Scalars['String']['output'];
};
// ...
export type InstanceNode = Node & {
  __typename?: 'InstanceNode';
  brands: BrandNodeConnection;
  cartSet: CartNodeConnection;
  cartitemSet: ItemNodeConnection;
  claimSet: GlobalDataNodeConnection;
  code?: Maybe<Scalars['String']['output']>;
  companytypes: CompanyTypeNodeConnection;
  contacts: ContactNodeConnection;
  documents: DocumentNodeConnection;
  downloadlogSet: DownloadLogNodeConnection;
  emailsubscriberSet: Array<EmailSubscriberType>;
  facebookUrl: Scalars['String']['output'];
  faqs: FaqNodeConnection;
  flatpageSet: FlatpageNodeConnection;
  homesliderSet: HomeSliderNodeConnection;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instagramUrl: Scalars['String']['output'];
  instance: ServiceContactNodeConnection;
  jobApplications: JobApplicationNodeConnection;
  linkedinUrl: Scalars['String']['output'];
  logo?: Maybe<Scalars['String']['output']>;
  markets: MarketNodeConnection;
  name: Scalars['String']['output'];
  orderitemSet: OrderItemNodeConnection;
  orders: OrderNodeConnection;
  posts: PostNodeConnection;
  products: ProductNodeConnection;
  projects: ProjectNodeConnection;
  projectsInstance: PromotedProjectNodeConnection;
  /** Please insert the emails separated for commas used in samples request in form No! shopping cart */
  sampleEmails?: Maybe<Scalars['String']['output']>;
  samplescontact: SampleContactNodeConnection;
  services: ServiceNodeConnection;
  stores: StoreNodeConnection;
  storesInstance: OfficeNodeConnection;
  twitterUrl: Scalars['String']['output'];
  userProfiles: ProfileNodeConnection;
  users: UserNodeConnection;
  vacancies: VacantNodeConnection;
};
// ...
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type OrderItemNode = Node & {
  __typename?: 'OrderItemNode';
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instance?: Maybe<InstanceNode>;
  objectId: Scalars['Int']['output'];
  order: OrderNode;
  pattern?: Maybe<ProductNode>;
  product?: Maybe<ProductNode>;
  quantity: Scalars['Int']['output'];
  reference?: Maybe<PatternNode>;
  unitPrice: Scalars['Float']['output'];
};

export type OrderItemNodeConnection = {
  __typename?: 'OrderItemNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<OrderItemNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};

/** A Relay edge containing a `OrderItemNode` and its cursor. */
export type OrderItemNodeEdge = {
  __typename?: 'OrderItemNodeEdge';
  /** A cursor for use in pagination */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge */
  node?: Maybe<OrderItemNode>;
};

export type OrderNode = Node & {
  __typename?: 'OrderNode';
  altName?: Maybe<Scalars['String']['output']>;
  created: Scalars['DateTime']['output'];
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instance: InstanceNode;
  items?: Maybe<OrderItemNodeConnection>;
  /** Notes about your order, e.g. special notes for delivery */
  notes: Scalars['String']['output'];
  shippingAddress?: Maybe<ShippingAddressNode>;
  status: OrderStatus;
  updated: Scalars['DateTime']['output'];
  user: UserNode;
};
// ...
export enum OrderStatus {
  /** New */
  A_1 = 'A_1',
  /** Processing */
  A_2 = 'A_2',
  /** Delivered */
  A_3 = 'A_3'
}
// ...
export type ProductNode = Node & {
  __typename?: 'ProductNode';
  claimSet: GlobalDataNodeConnection;
  code: Scalars['String']['output'];
  color: Scalars['String']['output'];
  /** Set as default product for its pattern (will be, for example, the default image appearing in listings) */
  default: Scalars['Boolean']['output'];
  /** Used to identify if there are available samples of this product */
  hasStock: Scalars['Boolean']['output'];
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  /** size 1280px X 800px, please upload the image in format PNG */
  image?: Maybe<Scalars['String']['output']>;
  imageForProduction?: Maybe<Scalars['String']['output']>;
  imageThumbLarge?: Maybe<Scalars['String']['output']>;
  imageThumbMedium?: Maybe<Scalars['String']['output']>;
  imageThumbSmall?: Maybe<Scalars['String']['output']>;
  instances: InstanceNodeConnection;
  /** Used to identify if this product is discontinued */
  isDiscontinued: Scalars['Boolean']['output'];
  /** Used to identify if this product is new */
  isNew: Scalars['Boolean']['output'];
  pantone?: Maybe<Scalars['String']['output']>;
  pattern: PatternNode;
  patternClass?: Maybe<ProductPatternClass>;
  predominantColors: Array<Scalars['String']['output']>;
  productonquickshipSet: ProductOnQuickshipNodeConnection;
  samples: SampleNodeConnection;
  /** Used to identify if Spradling is currently making samples of this product */
  samplesAvailable: Scalars['Boolean']['output'];
  stock: Scalars['String']['output'];
};
// ...
export type Query = {
// ...
  order?: Maybe<OrderNode>;
// ...
};
// ...
export type ShippingAddressNode = {
  __typename?: 'ShippingAddressNode';
  address: Scalars['String']['output'];
  city?: Maybe<CityNode>;
  cityName?: Maybe<Scalars['String']['output']>;
  company: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  lastName: Scalars['String']['output'];
  /** Indicates whether this is the main/default shipping address. */
  mainAddress: Scalars['Boolean']['output'];
  order: OrderNode;
  source: Scalars['String']['output'];
  zipCode?: Maybe<Scalars['Int']['output']>;
};
// ...
export type UserAddressNodeConnection = {
  __typename?: 'UserAddressNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<UserAddressNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};

/** A Relay edge containing a `UserAddressNode` and its cursor. */
export type UserAddressNodeEdge = {
  __typename?: 'UserAddressNodeEdge';
  /** A cursor for use in pagination */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge */
  node?: Maybe<UserAddressNode>;
};

export type UserNode = Node & {
  __typename?: 'UserNode';
  addresses: UserAddressNodeConnection;
  catalogues: CatalogueNodeConnection;
  createdPrivacyPolicies: PrivacyPolicyNodeConnection;
  dateJoined: Scalars['DateTime']['output'];
  email: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  fullName?: Maybe<Scalars['String']['output']>;
  groups?: Maybe<Array<Maybe<GroupNode>>>;
  /** Indica si el usuario tiene acceso a el boton de crear proyectos en library.spradling.group. */
  hasCreateProjectAccess: Scalars['Boolean']['output'];
  /** Indica si el usuario tiene acceso a library.spradling.group. */
  hasLibraryAccess: Scalars['Boolean']['output'];
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instance?: Maybe<InstanceNode>;
  /** Designates whether this user should be treated as active. Unselect this instead of deleting accounts. */
  isActive: Scalars['Boolean']['output'];
  /** Designates whether the user can log into this admin site. */
  isStaff: Scalars['Boolean']['output'];
  /** Designates that this user has all permissions without explicitly assigning them. */
  isSuperuser: Scalars['Boolean']['output'];
  jobpositionSet: JobPositionNodeConnection;
  lastLogin?: Maybe<Scalars['DateTime']['output']>;
  lastName: Scalars['String']['output'];
  orders: OrderNodeConnection;
  password: Scalars['String']['output'];
  profile?: Maybe<ProfileNode>;
  projectcommentSet: ProjectCommentNodeConnection;
  promotedProjectApprovers: PromotedProjectNodeConnection;
  quotations: QuotationNodeConnection;
  samples: SampleNodeConnection;
  unitSystem?: Maybe<Scalars['Int']['output']>;
  wishlists: WishlistNodeConnection;
};
// ...
export type GetOrderQuery = { __typename?: 'Query', order?: { __typename?: 'OrderNode', id: string, status: Types.OrderStatus, notes: string, created: any, updated: any, user: { __typename?: 'UserNode', id: string }, shippingAddress?: { __typename?: 'ShippingAddressNode', id: string, address: string, mainAddress: boolean, source: string, city?: { __typename?: 'CityNode', id: string, name: string, country: { __typename?: 'CountryNode', id: string, name: string } } | null } | null, items?: { __typename?: 'OrderItemNodeConnection', edges: Array<{ __typename?: 'OrderItemNodeEdge', node?: { __typename?: 'OrderItemNode', id: string, pattern?: { __typename?: 'ProductNode', id: string, pattern: { __typename?: 'PatternNode', id: string, name: string, slug: string } } | null, product?: { __typename?: 'ProductNode', id: string, color: string, code: string, pattern: { __typename?: 'PatternNode', id: string, name: string, slug: string } } | null } | null } | null> } | null } | null };